+++ lib/zaikio/hub/specialist.rb
@@ -7,7 +7,8 @@ module Zaikio
       include_root_in_json :specialist
 
       # Attributes
-      attributes :name, :updated_at, :created_at, :site_id, :kind, :owner_id, :capabilities
+      attributes :name, :updated_at, :created_at, :site_id, :kind, :owner_id,
+                 :capabilities, :reference
 
       # Associations
       has_many :availabilities, class_name: "Zaikio::Hub::Availability",